e-commerce/pyproject.toml
[tool.poetry]
name = "highload-final"
version = "0.1.0"
description = ""
authors = ["Danila Khlebokazov <d_khlebokazov@kbtu.kz>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.10"
django-environ = "^0.11.2"
django-debug-toolbar = "^4.4.6"
pre-commit = "^4.0.1"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
redis = "^5.2.0"
django-prometheus = "^2.3.1"
celery = "^5.4.0"
django-cors-headers = "^4.6.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
